# docs/api/openapi.v1.json
{
  "components": {
    "schemas": {
...
      "CalibrationRunRequest": {
        "properties": {
          "case_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Case Ids",
            "type": "array"
          },
          "rule_pack_id": {
            "maxLength": 80,
            "minLength": 3,
            "title": "Rule Pack Id",
            "type": "string"
          }
        },
        "required": [
          "rule_pack_id"
        ],
        "title": "CalibrationRunRequest",
        "type": "object"
      },
      "CalibrationRunStatus": {
        "properties": {
          "calibration_run_id": {
            "title": "Calibration Run Id",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "created_by_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By User Id"
          },
          "gate_status": {
            "title": "Gate Status",
            "type": "string"
          },
          "lifecycle_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lifecycle Run Id"
          },
          "lifecycle_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lifecycle Status"
          },
          "metrics": {
            "additionalProperties": true,
            "default": {},
            "title": "Metrics",
            "type": "object"
          },
          "rule_pack_id": {
            "title": "Rule Pack Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "calibration_run_id",
          "rule_pack_id",
          "status",
          "gate_status",
          "created_at"
        ],
        "title": "CalibrationRunStatus",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "password": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "ReviewCase": {
        "properties": {
          "assigned_to_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To User Id"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "default": {},
            "title": "Payload",
            "type": "object"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "resource_id": {
            "title": "Resource Id",
            "type": "string"
          },
          "resource_type": {
            "title": "Resource Type",
            "type": "string"
          },
          "review_id": {
            "title": "Review Id",
            "type": "string"
          },
          "review_type": {
            "title": "Review Type",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "review_id",
          "review_type",
          "resource_type",
          "resource_id",
          "severity",
          "status",
          "reason",
          "created_at"
        ],
        "title": "ReviewCase",
        "type": "object"
      },
      "ReviewDecision": {
        "properties": {
          "comment": {
            "default": "",
            "title": "Comment",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "enum": [
              "confirmed",
              "request_revision",
              "reject_promotion",
              "approve_promotion"
            ],
            "title": "Decision",
            "type": "string"
          },
          "decision_id": {
            "title": "Decision Id",
            "type": "string"
          },
          "review_id": {
            "title": "Review Id",
            "type": "string"
          },
          "reviewer_user_id": {
            "title": "Reviewer User Id",
            "type": "string"
          }
        },
        "required": [
          "decision_id",
          "review_id",
          "reviewer_user_id",
          "decision",
          "created_at"
        ],
        "title": "ReviewDecision",
        "type": "object"
      },
      "ReviewDecisionRequest": {
        "properties": {
          "comment": {
            "default": "",
            "maxLength": 3000,
            "title": "Comment",
            "type": "string"
          },
          "decision": {
            "enum": [
              "confirmed",
              "request_revision",
              "reject_promotion",
              "approve_promotion"
            ],
            "title": "Decision",
            "type": "string"
          }
        },
        "required": [
          "decision"
        ],
        "title": "ReviewDecisionRequest",
        "type": "object"
      },
...
      "RulePackCreateRequest": {
        "properties": {
          "action_adapter_version": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Action Adapter Version",
            "type": "string"
          },
          "consistency_rule_version": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Consistency Rule Version",
            "type": "string"
          },
          "evidence_policy_version": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Evidence Policy Version",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "scoring_weights_version": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Scoring Weights Version",
            "type": "string"
          },
          "supersedes_rule_pack_id": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Rule Pack Id"
          },
          "version": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Version",
            "type": "string"
          },
          "war_room_rule_version": {
            "maxLength": 80,
            "minLength": 1,
            "title": "War Room Rule Version",
            "type": "string"
          }
        },
        "required": [
          "name",
          "version",
          "war_room_rule_version",
          "consistency_rule_version",
          "action_adapter_version",
          "scoring_weights_version",
          "evidence_policy_version"
        ],
        "title": "RulePackCreateRequest",
        "type": "object"
      },
      "RulePackManifest": {
        "properties": {
          "action_adapter_version": {
            "title": "Action Adapter Version",
            "type": "string"
          },
          "activated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activated At"
          },
          "consistency_rule_version": {
            "title": "Consistency Rule Version",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "creator_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creator User Id"
          },
          "evidence_policy_version": {
            "title": "Evidence Policy Version",
            "type": "string"
          },
          "manifest": {
            "additionalProperties": true,
            "title": "Manifest",
            "type": "object"
          },
          "manifest_hash": {
            "title": "Manifest Hash",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rule_pack_id": {
            "title": "Rule Pack Id",
            "type": "string"
          },
          "scoring_weights_version": {
            "title": "Scoring Weights Version",
            "type": "string"
          },
          "status": {
            "enum": [
              "draft",
              "candidate",
              "active",
              "retired"
            ],
            "title": "Status",
            "type": "string"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "supersedes_rule_pack_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Rule Pack Id"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "war_room_rule_version": {
            "title": "War Room Rule Version",
            "type": "string"
          }
        },
        "required": [
          "rule_pack_id",
          "name",
          "version",
          "war_room_rule_version",
          "consistency_rule_version",
          "action_adapter_version",
          "scoring_weights_version",
          "evidence_policy_version",
          "manifest",
          "manifest_hash",
          "status",
          "created_at"
        ],
        "title": "RulePackManifest",
        "type": "object"
      },
      "RulePackReviewRequest": {
        "properties": {
          "comment": {
            "default": "",
            "maxLength": 2000,
            "title": "Comment",
            "type": "string"
          }
        },
        "title": "RulePackReviewRequest",
        "type": "object"
      },
...
      "SessionStatus": {
        "properties": {
          "absolute_expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Absolute Expires At"
          },
          "authenticated": {
            "title": "Authenticated",
            "type": "boolean"
          },
          "csrf_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Csrf Token"
          },
          "idle_expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idle Expires At"
          },
          "user": {
            "$ref": "#/components/schemas/UserIdentity"
          }
        },
        "required": [
          "authenticated",
          "user"
        ],
        "title": "SessionStatus",
        "type": "object"
      },
...
      "TrustSummary": {
        "properties": {
          "agent_admission_matrix": {
            "additionalProperties": true,
            "title": "Agent Admission Matrix",
            "type": "object"
          },
          "calibration": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CalibrationRunStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "calibration_status": {
            "title": "Calibration Status",
            "type": "string"
          },
          "data_coverage": {
            "title": "Data Coverage",
            "type": "number"
          },
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "golden_gate": {
            "additionalProperties": true,
            "title": "Golden Gate",
            "type": "object"
          },
          "pending_review_count": {
            "title": "Pending Review Count",
            "type": "integer"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "report_allowed": {
            "title": "Report Allowed",
            "type": "boolean"
          },
          "reviews": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ReviewCase"
            },
            "title": "Reviews",
            "type": "array"
          },
          "rule_pack": {
            "$ref": "#/components/schemas/RulePackManifest"
          },
          "rule_pack_history": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RulePackManifest"
            },
            "title": "Rule Pack History",
            "type": "array"
          },
          "security_gate": {
            "additionalProperties": true,
            "title": "Security Gate",
            "type": "object"
          }
        },
        "required": [
          "project_id",
          "rule_pack",
          "calibration_status",
          "golden_gate",
          "security_gate",
          "agent_admission_matrix",
          "data_coverage",
          "pending_review_count",
          "report_allowed",
          "generated_at"
        ],
        "title": "TrustSummary",
        "type": "object"
      },
      "UserIdentity": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "role": {
            "enum": [
              "admin",
              "analyst",
              "reviewer",
              "viewer"
            ],
            "title": "Role",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "username",
          "display_name",
          "role"
        ],
        "title": "UserIdentity",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "全球多源综合风险指数与世界局势情景模拟器",
    "title": "WorldPulse",
    "version": "1.2.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v3/auth/login": {
      "post": {
        "operationId": "local_login_api_v3_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Local Login"
      }
    },
...
    "/api/v3/calibration/runs": {
      "post": {
        "operationId": "calibration_run_create_api_v3_calibration_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationRunStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calibration Run Create"
      }
    },
    "/api/v3/calibration/runs/{calibration_run_id}": {
      "get": {
        "operationId": "calibration_run_detail_api_v3_calibration_runs__calibration_run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "calibration_run_id",
            "required": true,
            "schema": {
              "title": "Calibration Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationRunStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calibration Run Detail"
      }
    },
    "/api/v3/projects/{project_id}/trust-summary": {
      "get": {
        "operationId": "trust_summary_api_v3_projects__project_id__trust_summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/TrustSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trust Summary"
      }
    },
    "/api/v3/reviews": {
      "get": {
        "operationId": "review_list_api_v3_reviews_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewCase"
                  },
                  "title": "Response Review List Api V3 Reviews Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review List"
      }
    },
    "/api/v3/reviews/{review_id}": {
      "get": {
        "operationId": "review_detail_api_v3_reviews__review_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewCase"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Detail"
      }
    },
    "/api/v3/reviews/{review_id}/decision": {
      "post": {
        "operationId": "review_decision_api_v3_reviews__review_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDecision"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Decision"
      }
    },
    "/api/v3/rule-packs": {
      "get": {
        "operationId": "rule_pack_list_api_v3_rule_packs_get",
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RulePackManifest"
                  },
                  "title": "Response Rule Pack List Api V3 Rule Packs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Rule Pack List"
      },
      "post": {
        "operationId": "rule_pack_create_api_v3_rule_packs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RulePackCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/RulePackManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rule Pack Create"
      }
    },
    "/api/v3/rule-packs/{rule_pack_id}": {
      "get": {
        "operationId": "rule_pack_detail_api_v3_rule_packs__rule_pack_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "rule_pack_id",
            "required": true,
            "schema": {
              "title": "Rule Pack Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/RulePackManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rule Pack Detail"
      }
    },
    "/api/v3/rule-packs/{rule_pack_id}/activate": {
      "post": {
        "operationId": "rule_pack_activate_api_v3_rule_packs__rule_pack_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_pack_id",
            "required": true,
            "schema": {
              "title": "Rule Pack Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/RulePackManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rule Pack Activate"
      }
    },
    "/api/v3/rule-packs/{rule_pack_id}/approve": {
      "post": {
        "operationId": "rule_pack_approve_api_v3_rule_packs__rule_pack_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_pack_id",
            "required": true,
            "schema": {
              "title": "Rule Pack Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RulePackReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/RulePackManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rule Pack Approve"
      }
    },
    "/api/v3/rule-packs/{rule_pack_id}/submit": {
      "post": {
        "operationId": "rule_pack_submit_api_v3_rule_packs__rule_pack_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_pack_id",
            "required": true,
            "schema": {
              "title": "Rule Pack Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/RulePackManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rule Pack Submit"
      }
    },
...
  }
}